refactor(pythonista): extract MergerUI init/state/controls mixins
Continue REPOGROUND-LEGACY-RECONCILIATION-V1-T004 by moving remaining MergerUI
methods out of build.py into focused mixins. build.py becomes a thin composition
shell. Dual package/flat import wiring is preserved.
